+++ src/xmat-models/xmat-element-ref.model.ts
@@ -1,0 +1,11 @@
+import {ElementRef} from "@angular/core";
+
+export class XmatElementRef<T = any> extends ElementRef {
+
+    public nativeElement: T;
+
+    constructor(nativeElement: T) {
+        super(nativeElement);
+    }
+
+}