+++ changelog.md
@@ -265,3 +265,8 @@
 
 #1.3.12
 * Added new validations
+
+#1.3.13
+* New xmat-badge directive, cloned from material 6
+* Copy directives SCSS files to output allowing the use of xmat-library.scss
+* Changed xmat-action-text to directive, to allow using on components